+++ features/suite_test.go
@@ -613,6 +613,7 @@ func initializeScenario(sc *godog.ScenarioContext) {
 	initializeTasksSteps(sc)
 	initializeToolSteps(sc)
 	initializeAnswerSteps(sc)
+	initializeTaskWordSteps(sc)
 	initializeVersionSteps(sc)
 	initializeWorkSteps(sc)
 	initializeWorkMaterialSteps(sc)